refactor: extract top bar
